# docs/swagger.yaml
definitions:
  dto.ArAssetsResponse:
    properties:
      id:
        example: id
        type: string
      qrcode_image_path:
        example: url
        type: string
      speaking_audio_path:
        example: url
        type: string
      speaking_description:
        example: description
        type: string
      three_dimentional_path:
        example: url
        type: string
    type: object
  dto.BusinessCardBackgroundResponse:
    properties:
      business_card_background_color:
        example: '#ffffff'
        type: string
      business_card_background_image:
        example: url
        type: string
      id:
        type: string
    type: object
...
  dto.BusinessCardResponse:
    properties:
      accessCount:
        type: integer
      address:
        type: string
      businessCardBackgroundColor:
        description: background
        example: '#ffffff'
        type: string
      businessCardBackgroundImage:
        example: url
        type: string
      businessCardName:
        type: string
      businessCardPartsCoordinate:
        description: business card
        type: string
      companyName:
        type: string
      department:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
      officialPosition:
        type: string
      phoneNumber:
        type: string
      postalCode:
        type: string
      speakingAudioPath:
        type: string
      speakingDescription:
        type: string
      threeDimentionalModel:
        description: ar assets
        type: string
    type: object
  dto.CreateBackgroundRequest:
    properties:
      business_card_background:
        example: '#ffffff'
        type: string
      business_card_background_image:
        example: url
        type: string
        x-nullable: true
    type: object
  dto.CreateBusinessCardRequest:
    properties:
      address:
        example: 東京都渋谷区神南1-1-1
        type: string
        x-nullable: true
      ar_assets_id:
        description: ar assets
        example: ar_assets_id
        type: string
        x-nullable: true
      business_card_background_id:
        description: background
        example: id
        type: string
        x-nullable: true
      business_card_name:
        example: 会社
        type: string
        x-nullable: true
      business_card_parts_coordinate:
        description: business card
        example: id
        type: string
        x-nullable: true
      company_name:
        example: 会社名
        type: string
        x-nullable: true
      department:
        example: 部署
        type: string
        x-nullable: true
      display_name:
        example: 名前
        type: string
        x-nullable: true
      email:
        example: sample@example.com
        type: string
        x-nullable: true
      official_position:
        example: 役職
        type: string
        x-nullable: true
      phone_number:
        example: 090-1234-5678
        type: string
        x-nullable: true
      postal_code:
        example: 123-4567
        type: string
        x-nullable: true
      qrcode_image_path:
        example: url
        type: string
        x-nullable: true
    type: object
...
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AIRship API
  version: "1.0"
paths:
  /ar_assets/{ar_assets_id}:
    get:
      parameters:
      - description: ArAssets ID
        in: path
        name: ar_assets_id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ArAssetsResponse'
      tags:
      - ArAssets
...
  /users/ar_assets:
    get:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ArAssetsResponse'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - ArAssets
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: Image file to be uploaded
        in: formData
        name: qrcodeImage
        required: true
        type: file
      - example: description
        in: formData
        name: speaking_description
        type: string
      - example: url
        in: formData
        name: three_dimentional_ID
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ArAssetsResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - ArAssets
  /users/ar_assets/{ar_assets_id}:
    delete:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: ArAssets ID
        in: path
        name: ar_assets_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      tags:
      - ArAssets
    get:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: ArAssets ID
        in: path
        name: ar_assets_id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ArAssetsResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - ArAssets
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - example: description
        in: formData
        name: speaking_description
        type: string
      - example: url
        in: formData
        name: three_dimentional_ID
        type: string
      - description: ArAssets ID
        in: path
        name: ar_assets_id
        required: true
        type: string
      - description: Image file to be uploaded
        in: formData
        name: qrcodeIcon
        type: file
      responses:
        "201":
          description: Created
      security:
      - ApiKeyAuth: []
      tags:
      - ArAssets
  /users/business_card_backgrounds:
    get:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.BusinessCardBackgroundResponse'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - BusinessCardBackground
    post:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: BusinessCardBackground
        in: body
        name: dto.CreateBackgroundRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBackgroundRequest'
      responses:
        "201":
          description: Created
      security:
      - ApiKeyAuth: []
      tags:
      - BusinessCardBackground
  /users/business_cards:
    get:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.BusinessCardResponse'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - BusinessCard
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: Image file to be uploaded
        in: formData
        name: BusinessCardBackgroundImage
        required: true
        type: file
      - example: 東京都渋谷区神南1-1-1
        in: formData
        name: address
        type: string
        x-nullable: true
      - description: ar assets
        example: ar_assets_id
        in: formData
        name: ar_assets_id
        type: string
        x-nullable: true
      - description: background
        example: id
        in: formData
        name: business_card_background_id
        type: string
        x-nullable: true
      - example: 会社
        in: formData
        name: business_card_name
        type: string
        x-nullable: true
      - description: business card
        example: id
        in: formData
        name: business_card_parts_coordinate
        type: string
        x-nullable: true
      - example: 会社名
        in: formData
        name: company_name
        type: string
        x-nullable: true
      - example: 部署
        in: formData
        name: department
        type: string
        x-nullable: true
      - example: 名前
        in: formData
        name: display_name
        type: string
        x-nullable: true
      - example: sample@example.com
        in: formData
        name: email
        type: string
        x-nullable: true
      - example: 役職
        in: formData
        name: official_position
        type: string
        x-nullable: true
      - example: 090-1234-5678
        in: formData
        name: phone_number
        type: string
        x-nullable: true
      - example: 123-4567
        in: formData
        name: postal_code
        type: string
        x-nullable: true
      - example: url
        in: formData
        name: qrcode_image_path
        type: string
        x-nullable: true
      responses:
        "201":
          description: Created
      security:
      - ApiKeyAuth: []
      tags:
      - BusinessCard
  /users/business_cards/{business_card_id}:
    delete:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: BusinessCard ID
        in: path
        name: business_card_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      tags:
      - BusinessCard
    get:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: BusinessCard ID
        in: path
        name: business_card_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BusinessCardResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - BusinessCard
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: BusinessCard ID
        in: path
        name: business_card_id
        required: true
        type: string
      - description: Image file to be uploaded
        in: formData
        name: BusinessCardBackgroundImage
        required: true
        type: file
      - description: BusinessCard
        in: body
        name: CreateBusinessCardRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBusinessCardRequest'
      responses:
        "201":
          description: Created
      security:
      - ApiKeyAuth: []
      tags:
      - BusinessCard
...
swagger: "2.0"